requirements.txt
google-genai
requests
python-dotenv
fuzzywuzzy
python-Levenshtein
lxml